Changed how Theory is parsed.
