+++ MonoDevelop.XUnit/MonoDevelop.UnitTesting.XUnit/NUnitProjectTestSuite.cs
@@ -36,7 +36,6 @@ using MonoDevelop.Ide;
 using MonoDevelop.Ide.TypeSystem;
 using MonoDevelop.Projects;
 using MonoDevelop.UnitTesting.XUnit.External;
-using MonoDevelop.XUnit;
 using ProjectReference = MonoDevelop.Projects.ProjectReference;
 
 namespace MonoDevelop.UnitTesting.XUnit
