+++ XUnitRunner/XUnitRunner.cs
@@ -47,7 +47,7 @@ namespace XUnitRunner
 			Type t = Type.GetType("Mono.Runtime");
 			var conf = ConfigReader.Load(assembly);
 			if (t != null) {
-				//TODO support below
+				// TODO: support below
 				conf.PreEnumerateTheories = true;
 				conf.ParallelizeTestCollections = false;
 				conf.AppDomain = AppDomainSupport.Denied;
@@ -68,16 +68,15 @@ namespace XUnitRunner
 							Id = testCase.UniqueID,
 							Type = testCase.TestMethod.TestClass.Class.Name,
 							Method = testCase.TestMethod.Method.Name,
-							DisplayName = testCase.DisplayName,
 							Args = testCase.TestMethodArguments
 						});
 					}
 				}
 				// sort by type, method
 				infos.Sort((info1, info2) => {
-					int i = info1.Type.CompareTo(info2.Type);
+					int i = string.Compare(info1.Type, info2.Type, StringComparison.Ordinal);
 					if (i == 0)
-						i = info1.Method.CompareTo(info2.Method);
+						i = string.Compare(info1.Method, info2.Method, StringComparison.Ordinal);
 					return i;
 				});
 			}
@@ -109,7 +108,7 @@ namespace XUnitRunner
 					testInfo.Id = firstItem.Id;
 					testInfo.Type = firstItem.Type;
 					testInfo.Method = firstItem.Method;
-					testInfo.Name = firstItem.Name;
+					testInfo.Name = firstItem.DisplayName;
 					testInfo.Args = firstItem.Args;
 					return;
 				}
@@ -164,59 +163,6 @@ namespace XUnitRunner
 				controller.RunTests(discoveryVisitor.TestCases, executionVisitor, executionOptions);
 				executionVisitor.Finished.WaitOne();
 			}
-		}
-
-		class TestCaseInfo
-		{
-			public string Id;
-			public string Type;
-			public string Method;
-			public string DisplayName;
-			public object[] Args;
-
-			string name;
-			public string Name {
-				get {
-					if (name == null) {
-						if (DisplayName.StartsWith(Type + "." + Method))
-							name = Method;
-						else
-							name = DisplayName;
-					}
-					return name;
-				}
-			}
-
-			void parseName()
-			{
-				// this is [theory], xunit v2 where each theory is a separate test case
-				if (Args != null && Args.Length > 0) {
-					string[] typeParts = Type.Split('.');
-					nameParts = new string[typeParts.Length + 2];
-					typeParts.CopyTo(nameParts, 0);
-					var leftBracket = DisplayName.IndexOf('(');
-					var lastDot = DisplayName.LastIndexOf('.', leftBracket);
-					nameParts[typeParts.Length] = DisplayName.Substring(lastDot + 1);
-					nameParts[typeParts.Length + 1] = Method;
-				}
-				// this is [fact]
-				else {
-					string[] typeParts = Type.Split('.');
-					nameParts = new string[typeParts.Length + 1];
-					typeParts.CopyTo(nameParts, 0);
-					nameParts[typeParts.Length] = Method;
-				}
-			}
-
-			string[] nameParts;
-			public string[] NameParts {
-				get {
-					if (nameParts == null)
-						parseName();
-					return nameParts;
-				}
-			}
-
 		}
 	}
 }
