+++ XUnitRunner/XUnitRunner.csproj
@@ -61,6 +61,9 @@
     <Compile Include="..\MonoDevelop.XUnit\XUnitTestInfo.cs">
       <Link>XUnitTestInfo.cs</Link>
     </Compile>
+    <Compile Include="..\MonoDevelop.XUnit\UnitTestInfo.cs">
+      <Link>UnitTestInfo.cs</Link>
+    </Compile>
   </ItemGroup>
   <ItemGroup>
     <None Include="packages.config" />
